Filter all resources of nested projects from their parent
This way resources only show up in the project closest to them (shortest path to parent)
The "Open Resource" dialog no longer offers to open the same resources multiple times,
"File Search" no longer finds matches in the same file multiple times (in multiple projects)
This greatly improves the user experience for Gradle and Maven projects which normally lay out
their related projects in a hierarchial way.

Some usecases that relied on the previous behavior may degrade as a result,
which usecases these are and if the degradation should be considered a defect is yet to be determined.

Signed-off-by: Christoph Obexer <cobexer@gmail.com>